# services/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the account service for PayGidi.",
        "title": "PayGidi Account Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api.paygidi.site",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/otp/request/{otpType}": {
            "post": {
                "description": "Request an OTP to be sent via email or SMS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Request OTP",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OTP type (email or phone)",
                        "name": "otpType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "OTP request data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.RequestOTPDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/business/docs": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the verification documents for a business account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Update business documents",
                "parameters": [
                    {
                        "description": "Business documents data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UpdateBusinessDocsDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business documents updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/business/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the profile information for the authenticated business account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Get business profile",
                "responses": {
                    "200": {
                        "description": "Business profile retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the profile information for a business account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Update business profile",
                "parameters": [
                    {
                        "description": "Business profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UpdateBusinessProfileDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Business profile updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "validators.RequestOTPDto": {
            "type": "object",
            "required": [
                "forWhat"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "forWhat": {
                    "type": "string",
                    "enum": [
                        "completeRegister",
                        "login",
                        "resetPassword",
                        "twoFactorAuth"
                    ]
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10
                }
            }
        },
        "validators.UpdateBusinessDocsDto": {
            "type": "object",
            "properties": {
                "additionalDocs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "registrationDoc": {
                    "type": "string"
                }
            }
        },
        "validators.UpdateBusinessProfileDto": {
            "type": "object",
            "properties": {
                "businessType": {
                    "type": "string"
                },
                "facebook": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "linkedIn": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "registrationNumber": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5
                },
                "twitter": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
    }
}